FEAT: 게시판 추가

+++ src/main/java/com/luti/payment/dto/PaymentListResponseDTO.java
@@ -16,7 +16,7 @@ public class PaymentListResponseDTO {
 
     private Integer paymentNo;        // 결제 번호
     private Integer paymentCd;        // 결제 방식 코드
-    private Integer userId;           // 사용자 ID
+    private Long userId;           // 사용자 ID
     private Integer totalPrice;       // 총 결제 금액
     private Integer paymentState;     // 결제 상태
     private LocalDate paymentDate;    // 결제 일자
